V 1.2 BETA test 5

+++ main.py
@@ -10,11 +10,14 @@ GPIO.setmode(GPIO.BCM)
 GPIO.setup(3, GPIO.IN, pull_up_down=GPIO.PUD_UP)
 
 last_button_press_time = 0
+i = 0
 
 # Function to be executed when the button is pressed
 def button_callback(channel):
-
+    global i
     global last_button_press_time
+
+    print('i =', i)
 
     # Calculate time since last button press
     time_since_last_press = time.monotonic() - last_button_press_time
@@ -26,22 +29,27 @@ def button_callback(channel):
     last_button_press_time = time.monotonic()
 
     # Check if there is an active network connection
-    if connection_status():
+    if connection_status() and i == 0:
         print('CONECTION OK')
         # Check if ffmpeg process is currently running
         process = check_process('ffmpeg')
         
         if process:
             print('IF PROCESS')
+            i += 1
             # If ffmpeg process is currently running, kill it and wait for 10 seconds
             kill_process('ffmpeg')
             time.sleep(10)
 
         else:
             print('NOT PROCESS')
+            i += 1
             # If ffmpeg process is not running and there is an active network connection, start the ffmpeg process and wait for 10 seconds
             ffmpeg = start_ffmpeg()
             time.sleep(10)
+
+    else:
+        i = 0
 
 # Register callback function for event
 GPIO.add_event_detect(3, GPIO.FALLING, callback=button_callback, bouncetime=1000)
